+++ packages/strapi-design-system/src/ModalLayout/__tests__/ModalLayout.spec.js
@@ -462,9 +462,7 @@ describe('ModalLayout', () => {
         max-height: 60vh;
       }
 
-      <body
-        class="lock-body-scroll"
-      >
+      <body>
         <div>
           <div
             class="c0"
@@ -496,91 +494,93 @@ describe('ModalLayout', () => {
             class="c1"
           >
             <div>
-              <div
-                aria-labelledby="title"
-                aria-modal="true"
-                class="c2 c3 c4"
-                role="dialog"
-              >
+              <div>
                 <div
-                  class="c2 c5 c6"
+                  aria-labelledby="title"
+                  aria-modal="true"
+                  class="c2 c3 c4"
+                  role="dialog"
                 >
                   <div
-                    class="c2 c7"
-                  >
-                    Modal Title
-                    <button
-                      aria-disabled="false"
-                      aria-label="Close the modal"
-                      class="c8 c9"
-                      type="button"
-                    >
-                      <svg
-                        fill="none"
-                        height="1em"
-                        viewBox="0 0 24 24"
-                        width="1em"
-                        xmlns="http://www.w3.org/2000/svg"
-                      >
-                        <path
-                          d="M24 2.417L21.583 0 12 9.583 2.417 0 0 2.417 9.583 12 0 21.583 2.417 24 12 14.417 21.583 24 24 21.583 14.417 12 24 2.417z"
-                          fill="#212134"
-                        />
-                      </svg>
-                    </button>
-                  </div>
-                </div>
-                <div
-                  class="c2 c10 c11"
-                >
-                  Hello World
-                </div>
-                <div
-                  class="c2 c5 c12"
-                >
-                  <div
-                    class="c2 c7"
+                    class="c2 c5 c6"
                   >
                     <div
-                      class="c2 c13 c14"
+                      class="c2 c7"
                     >
+                      Modal Title
                       <button
                         aria-disabled="false"
-                        class="c8 c15"
+                        aria-label="Close the modal"
+                        class="c8 c9"
                         type="button"
                       >
-                        <span
-                          class="c16 c17"
+                        <svg
+                          fill="none"
+                          height="1em"
+                          viewBox="0 0 24 24"
+                          width="1em"
+                          xmlns="http://www.w3.org/2000/svg"
                         >
-                          Cancel
-                        </span>
+                          <path
+                            d="M24 2.417L21.583 0 12 9.583 2.417 0 0 2.417 9.583 12 0 21.583 2.417 24 12 14.417 21.583 24 24 21.583 14.417 12 24 2.417z"
+                            fill="#212134"
+                          />
+                        </svg>
                       </button>
                     </div>
+                  </div>
+                  <div
+                    class="c2 c10 c11"
+                  >
+                    Hello World
+                  </div>
+                  <div
+                    class="c2 c5 c12"
+                  >
                     <div
-                      class="c2 c13 c14"
+                      class="c2 c7"
                     >
-                      <button
-                        aria-disabled="false"
-                        class="c8 c18"
-                        type="button"
+                      <div
+                        class="c2 c13 c14"
                       >
-                        <span
-                          class="c16 c17"
+                        <button
+                          aria-disabled="false"
+                          class="c8 c15"
+                          type="button"
                         >
-                          Add new stuff
-                        </span>
-                      </button>
-                      <button
-                        aria-disabled="false"
-                        class="c8 c19"
-                        type="button"
+                          <span
+                            class="c16 c17"
+                          >
+                            Cancel
+                          </span>
+                        </button>
+                      </div>
+                      <div
+                        class="c2 c13 c14"
                       >
-                        <span
-                          class="c16 c17"
+                        <button
+                          aria-disabled="false"
+                          class="c8 c18"
+                          type="button"
                         >
-                          Finish
-                        </span>
-                      </button>
+                          <span
+                            class="c16 c17"
+                          >
+                            Add new stuff
+                          </span>
+                        </button>
+                        <button
+                          aria-disabled="false"
+                          class="c8 c19"
+                          type="button"
+                        >
+                          <span
+                            class="c16 c17"
+                          >
+                            Finish
+                          </span>
+                        </button>
+                      </div>
                     </div>
                   </div>
                 </div>
