+++ packages/strapi-design-system/src/index.js
@@ -14,6 +14,7 @@ export * from './CarouselInput';
 export * from './Checkbox';
 export * from './Combobox';
 export * from './Dialog';
+export * from './DismissableLayer';
 export * from './DatePicker';
 export * from './Divider';
 export * from './EmptyStateLayout';
